Conflicts:
	create.c
	dll.c
	implement.h
	pthread_win32_attach_detach_np.c
	ptw32_MCS_lock.c
	tests/Bmakefile
	tests/Wmakefile
	tests/exception2.c
	tests/exception3.c
	tests/mutex6n.c
	tests/once2.c
	tests/once3.c

+++ tests/openmp1.c
@@ -29,7 +29,7 @@ static void *_thread(void* Id) {
     if (Verbose && i%1000==0) {
       int tid = omp_get_thread_num();
 #  pragma omp critical
-      printf("thread %d : tid %d handles %d\n",(int)Id,tid,i);
+      printf("thread %d : tid %d handles %d\n",(int)(size_t)Id,tid,i);
     }
 #endif
 
@@ -44,9 +44,9 @@ static void *_thread(void* Id) {
 #ifdef _OPENMP
 #  pragma omp critical
 #endif
-    printf("Id %d : %s : %d(should be %d)\n",(int)Id, __FUNCTION__, Sum,ShouldSum);
+    printf("Id %d : %s : %d(should be %d)\n",(int)(size_t)Id, __FUNCTION__, Sum,ShouldSum);
   }
-  if (Sum == ShouldSum) ThreadOK[(int)Id] = 1;
+  if (Sum == ShouldSum) ThreadOK[(int)(size_t)Id] = 1;
   return NULL;
 }
 
@@ -62,7 +62,7 @@ static void MainThread() {
       int tid = omp_get_thread_num();
 #  pragma omp critical
       printf("Main : tid %d\n",tid);
-      _thread((void *)tid);
+      _thread((void *)(size_t)tid);
 #endif
   }
   return;
